requirements.txt
# requirements for HPC, Redis, and Flask


fastapi
uvicorn
opencv-python
ultralytics
numpy
pandas
redis
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
aiortc